requirements.txt
ultralytics
iopaint
opencv-python
numpy
Pillow
torch
torchvision
